Rename to correspond to definitions
