+++ Epsilon.Host.Frontend/src/components/CompetenceProfileLegend.vue
